+ Extracted insight section from home
